pyproject.toml
[tool.poetry]
name = "board-for-sky"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
pillow = "^11.1.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"
djangorestframework-simplejwt = "^5.4.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
